requirements.txt
Flask
scikit-learn
numpy
pillow